backend/requirements.txt
flask==3.1.2
flask-cors==6.0.2
flask-sqlalchemy==3.1.1
pymysql==1.1.2
redis==7.0.1
neo4j==5.28.3
openai==2.20.0
python-dotenv==1.2.1
